src/uab/pyproject.toml
[project]
name = "uab"
version = "0.1.0"
description = "Universal Asset Browser"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyside6>=6.10.1",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "openexr>=3.2.0",
    "imageio>=2.31.0",
    "imath>=0.0.2",
    "opencv-python>=4.11.0.86",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { uab = "." }
packages = ["uab", "uab.core", "uab.ui"]
